Edición y borrado de marcas

+++ app/Http/Controllers/MarcaController.php
@@ -11,10 +11,22 @@ class MarcaController extends Controller
 {
     //
 
+    public function editar($id=null){
+        $marca = null;
+        if($id){
+            $marca = Marca::find($id);
+            if(!$marca || ($marca && $marca->tienda_id != Auth::id())){
+                return back()->withErrors(["warning"=>"La marca solicitada no existe"]);
+            }
+        }
+
+        return view('marcas.editar',["marca"=>$marca]);
+    }
+
     public function listar(){
 
         //Marcas de la tienda
-        $marcasTienda = Marca::where('tienda_id',Auth::id())->get();
+        $marcasTienda = Marca::with('productos')->where('tienda_id',Auth::id())->get();
         //Marcas generales
         $marcasYeelu = Marca::where('tienda_id',null)->get();
 
@@ -24,7 +36,8 @@ class MarcaController extends Controller
     public function guardar(Request $request){
         $valido = $request->validate([
             "nombre" => ["required","string","max:150"],
-            "foto" => ["image","nullable","max:8192"]
+            "foto" => ["image","nullable","max:8192"],
+            "id" => ["sometimes","required","exists:marcas,id"]
         ],[
             'nombre.required' => 'El nombre es obligatorio',
             'nombre.string' => 'El nombre debe ser un texto válido',
@@ -33,8 +46,16 @@ class MarcaController extends Controller
             'foto.max' => 'El tamaño máximo de la foto es de 8MB'
         ]);
         try{
-            $marca = new Marca();
+            $marca = null;
+            if(isset($valido['id'])){
+                $marca = Marca::find($valido['id']);
+            }else{
+                $marca = new Marca();
+            }
             $marca->nombre = $request->nombre;
+            if(isset($valido['foto']) && $marca->foto && Storage::disk('public')->exists($marca->foto)){
+                Storage::disk('public')->delete($marca->foto);
+            }
             if(isset($valido['foto'])){
                 $marca->foto = $request->file('foto')->store('marcas','public');
             }
@@ -43,12 +64,28 @@ class MarcaController extends Controller
             $marca->save();
             return redirect('marcas')->withErrors(['success' => "Marca guardada"]);
         }catch(\Exception $e){
-            if (Storage::disk('public')->exists($marca->foto)) {
+            if (!isset($valido['id']) && Storage::disk('public')->exists($marca->foto)) {
                 Storage::disk('public')->delete($marca->foto);
             }
             return back()->withErrors([
                 'danger' => 'Se ha producido un error en el sistema.'.(config('app.env')!="production" ? " ".$e->getMessage():"")
             ]);
+        }
+    }
+
+    public function borrar($id){
+        $marca = Marca::find($id);
+        if(!$marca || ($marca && $marca->tienda_id != Auth::id())){
+            return back()->withErrors(["warning"=>"La marca solicitada no existe"]);
+        }
+        else if(count($marca->productos)){
+            return back()->withErrors(["danger"=>"No puedes borrar una marca con productos asociados"]);
+        }else{
+            if ($marca->foto && Storage::disk('public')->exists($marca->foto)) {
+                Storage::disk('public')->delete($marca->foto);
+            }
+            $marca->delete();
+            return redirect('marcas')->withErrors(["success"=>"Marca borrada"]);
         }
     }
 }
